requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
